+++ src/anemoi/transform/variables/variables.py
@@ -7,12 +7,21 @@
 # granted to it by virtue of its status as an intergovernmental organisation
 # nor does it submit to any jurisdiction.
 
-
+import logging
+from collections.abc import Sized
+from typing import TYPE_CHECKING
 from typing import Any
 from typing import Dict
 from typing import Union
 
+from anemoi.utils.dates import as_timedelta
+
 from anemoi.transform.variables import Variable
+
+if TYPE_CHECKING:
+    from datetime import timedelta
+
+LOG = logging.getLogger(__name__)
 
 
 class VariableFromMarsVocabulary(Variable):
@@ -81,6 +90,22 @@ class VariableFromMarsVocabulary(Variable):
     def time_processing(self):
         """Get the time processing type of the variable."""
         return self.data.get("process")
+
+    @property
+    def period(self) -> Union["timedelta", None]:
+        """Get the variable's period as a timedelta.
+        For instantaneous variables, returns a timedelta of 0. For non-instantaneous variables, returns `None` if this information is missing.
+        """
+        if self.is_instantanous:
+            return as_timedelta(0)
+
+        if not (period := self.data.get("period")):
+            return None
+
+        if not isinstance(period, Sized) or len(period) != 2:
+            return None
+
+        return as_timedelta(period[1]) - as_timedelta(period[0])
 
     @property
     def grib_keys(self) -> Dict[str, Any]: